Add sample method to SearchNode
Note that this is still a work in progress. search, selectAction, and
playout still need to be implemented. Furthermore, there's still an open
question on how the children/ancestry should be represented inside of
SearchNode. A fixed array is rather inefficient since it requires
a number of elements equal to max{|actions|,|percepts| for all
environments E}, regardless of which one we actually choose at runtime.

A future implementation of search just needs to run sample repeatedly
since the best-guess (action with maximum m_mean) gets asymptotically
better each time sample is run. How should we determine when to stop,
seeing as we don't have a concept of "limited" computation time (insofar
as that we actually need to be able to run it in a timely manner so that
we can get nice results for the report).

Squashed commit of the following:

commit 17da37d95af15979581125088c9cf18b73707a85
Author: Andrew Haigh <nelfin@gmail.com>
Date:   Tue Oct 9 19:11:00 2012 +1100

    Add stubs to genPercept* in Agent

commit 339aa621f5c16e6d14d6755c48a1ecdd59cca45c
Author: Andrew Haigh <nelfin@gmail.com>
Date:   Tue Oct 9 19:03:49 2012 +1100

    Fix in (de/con)-structor for SearchNode

commit 6435a10cd9ec48a3e2111169f8fe3d33ae562a25
Author: Andrew Haigh <nelfin@gmail.com>
Date:   Tue Oct 9 18:58:14 2012 +1100

    Change signature of genPercept* in Agent

commit 0f938784e51bcd4cc00f5555f52b6f43cf47302e
Author: Andrew Haigh <nelfin@gmail.com>
Date:   Mon Oct 8 17:37:56 2012 +1100

    Fix sample on a chance node to reference child

commit e71d2c64ecc3ff3b8d85143bc9aeb77e4492eb73
Author: Andrew Haigh <nelfin@gmail.com>
Date:   Mon Oct 8 17:08:26 2012 +1100

    Add UCT (Monte-Carlo with UCB policy)

    WIP: halfway there, please don't merge

commit bc8cf1a7c74af2cc7cf13dd83d347061c7eadf0a
Author: Andrew Haigh <nelfin@gmail.com>
Date:   Sun Oct 7 18:54:57 2012 +1100

    WIP: add sample method to SearchNode