Add simulations member to agent
I don't know why this was missing. Mayank pointed me to the section of
the readme that detailed the common option mc-simulations, but none of
the conf files specify it, so it may be a problem initially. Luckily,
however, nothing actually uses search(), so we don't run into too much
of an issue with the code in its current state.