Add search method to search.cpp
And that's all folks. The last unimplemented method in rho-UCT.
Theoretically, it should be able to learn now. I have a feeling that it
doesn't though, I just tried it on the coinflip environment and after
some fiddling, i.e. we need to add mc-simulations to the *.conf files (I
went with 4 because it's sloooooooow right now. I should probably look
into that), it was only getting an average reward of 0.03 after 1024
cycles. Surely it can do better than that.