Composite can now change between many environments
The composite environment can now accept a sequence of environments. At
present this is capped at 10 for simplicity. The environmentCode and
startCode functions could probably be nicer. The config file
demonstrates that the Composite environment can start with an
environment, transfer to a different one and then transfer back - which
is what the assignment says to do.